+++ pkg/pillar/pubsub/pubsub_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) 2019-2020 Zededa, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
 package pubsub
 
 import (
@@ -21,48 +24,6 @@ var (
 		aString: "aString",
 	}
 )
-
-type EmptyDriver struct{}
-
-func (e *EmptyDriver) Publisher(global bool, name, topic string, persistent bool, updaterList *Updaters, restarted Restarted, differ Differ) (DriverPublisher, error) {
-	return &EmptyDriverPublisher{}, nil
-}
-func (e *EmptyDriver) Subscriber(global bool, name, topic string, persistent bool, C chan Change) (DriverSubscriber, error) {
-	return &EmptyDriverSubscriber{}, nil
-}
-func (e *EmptyDriver) DefaultName() string {
-	return "empty"
-}
-
-type EmptyDriverPublisher struct{}
-
-func (e *EmptyDriverPublisher) Start() error {
-	return nil
-}
-func (e *EmptyDriverPublisher) Load() (map[string][]byte, bool, error) {
-	return make(map[string][]byte), false, nil
-}
-func (e *EmptyDriverPublisher) Publish(key string, item []byte) error {
-	return nil
-}
-func (e *EmptyDriverPublisher) Unpublish(key string) error {
-	return nil
-}
-func (e *EmptyDriverPublisher) Restart(restarted bool) error {
-	return nil
-}
-
-type EmptyDriverSubscriber struct{}
-
-func (e *EmptyDriverSubscriber) Start() error {
-	return nil
-}
-
-// Load load entire persisted data set into a map
-func (e *EmptyDriverSubscriber) Load() (map[string][]byte, bool, error) {
-	res := make(map[string][]byte)
-	return res, false, nil
-}
 
 func TestHandleModify(t *testing.T) {
 	ps := New(&EmptyDriver{})
